Move ChannelList, UserStore into client crate
