+++ crates/client/Cargo.toml
@@ -10,10 +10,13 @@ test-support = []
 gpui = { path = "../gpui" }
 util = { path = "../util" }
 rpc = { path = "../rpc" }
+sum_tree = { path = "../sum_tree" }
 
 anyhow = "1.0.38"
 async-recursion = "0.3"
 async-tungstenite = { version = "0.14", features = ["async-tls"] }
+futures = "0.3"
+image = "0.23"
 lazy_static = "1.4.0"
 log = "0.4"
 parking_lot = "0.11.1"
@@ -22,4 +25,5 @@ rand = "0.8.3"
 smol = "1.2.5"
 surf = "2.2"
 thiserror = "1.0.29"
+time = "0.3"
 tiny_http = "0.8"
