+++ crates/zed/Cargo.toml
@@ -68,7 +68,7 @@ smol = "1.2.5"
 surf = "2.2"
 tempdir = { version = "0.3.7", optional = true }
 thiserror = "1.0.29"
-time = { version = "0.3" }
+time = "0.3"
 tiny_http = "0.8"
 toml = "0.5"
 tree-sitter = "0.19.5"
