+++ crates/zed/src/test.rs
@@ -1,20 +1,15 @@
 use crate::{
     assets::Assets,
-    channel::ChannelList,
-    http::{HttpClient, Request, Response, ServerResponse},
     language,
     settings::{self, ThemeRegistry},
-    user::UserStore,
     AppState,
 };
-use anyhow::Result;
 use buffer::LanguageRegistry;
-use client::Client;
-use futures::{future::BoxFuture, Future};
+use client::{http::ServerResponse, test::FakeHttpClient, ChannelList, Client, UserStore};
 use gpui::MutableAppContext;
 use parking_lot::Mutex;
 use project::fs::FakeFs;
-use std::{fmt, sync::Arc};
+use std::sync::Arc;
 
 #[cfg(test)]
 #[ctor::ctor]
@@ -40,34 +35,4 @@ pub fn test_app_state(cx: &mut MutableAppContext) -> Arc<AppState> {
         user_store,
         fs: Arc::new(FakeFs::new()),
     })
-}
-
-pub struct FakeHttpClient {
-    handler:
-        Box<dyn 'static + Send + Sync + Fn(Request) -> BoxFuture<'static, Result<ServerResponse>>>,
-}
-
-impl FakeHttpClient {
-    pub fn new<Fut, F>(handler: F) -> Arc<dyn HttpClient>
-    where
-        Fut: 'static + Send + Future<Output = Result<ServerResponse>>,
-        F: 'static + Send + Sync + Fn(Request) -> Fut,
-    {
-        Arc::new(Self {
-            handler: Box::new(move |req| Box::pin(handler(req))),
-        })
-    }
-}
-
-impl fmt::Debug for FakeHttpClient {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        f.debug_struct("FakeHttpClient").finish()
-    }
-}
-
-impl HttpClient for FakeHttpClient {
-    fn send<'a>(&'a self, req: Request) -> BoxFuture<'a, Result<Response>> {
-        let future = (self.handler)(req);
-        Box::pin(async move { future.await.map(Into::into) })
-    }
 }
